Shared wrapper traits bip44/cip1852
