+++ app/api/ada/lib/storage/models/Bip44Wallet/interfaces.js
@@ -58,83 +58,7 @@ export interface IBip44Wallet {
   getWrapperId(): number;
   getPublicDeriverLevel(): number;
   getSigningLevel(): number | null;
-  getPrivateDeriverLevel(): number | null;
-  getPrivateDeriverKeyDerivationId(): number | null;
   getProtocolMagic(): number;
-}
-
-export type IDerivePublicFromPrivateRequest = {|
-  publicDeriverMeta: {|
-    name: string,
-  |},
-  decryptPrivateDeriverPassword: null | string,
-  /**
-   * void -> do not store key
-   * null -> store as unencrypted
-   * string  -> store and encrypt
-   */
-  encryptPublicDeriverPassword?: null | string,
-  initialDerivations: TreeInsert<any>,
-  path: Array<number>,
-|};
-export type IDerivePublicFromPrivateResponse<Row> = AddPublicDeriverResponse<Row>;
-export type IDerivePublicFromPrivateFunc<Row> = (
-  body: IDerivePublicFromPrivateRequest
-) => Promise<IDerivePublicFromPrivateResponse<Row>>;
-export interface IDerivePublicFromPrivate {
-  +rawDerivePublicDeriverFromPrivate: RawTableVariation<
-    IDerivePublicFromPrivateFunc<mixed>,
-    {|
-      DerivePublicDeriverFromKey: Class<DerivePublicDeriverFromKey>,
-    |},
-    IDerivePublicFromPrivateRequest
-  >;
-  +derivePublicDeriverFromPrivate: IDerivePublicFromPrivateFunc<mixed>;
-}
-
-export type IGetPrivateDeriverKeyRequest = void;
-export type IGetPrivateDeriverKeyResponse = {
-  keyRow: $ReadOnly<KeyRow>,
-  keyDerivation: $ReadOnly<KeyDerivationRow>,
-};
-export type IGetPrivateDeriverKeyFunc = (
-  body: IGetPrivateDeriverKeyRequest
-) => Promise<IGetPrivateDeriverKeyResponse>;
-export interface IGetPrivateDeriverKey {
-  +rawGetPrivateDeriverKey: RawVariation<
-    IGetPrivateDeriverKeyFunc,
-    {|
-      GetKeyForDerivation: Class<GetKeyForDerivation>,
-    |},
-    IGetPrivateDeriverKeyRequest
-  >;
-  +getPrivateDeriverKey: IGetPrivateDeriverKeyFunc;
-
-  +rawChangePrivateDeriverPassword: RawVariation<
-    IChangePasswordRequestFunc,
-    {|
-      GetKeyForDerivation: Class<GetKeyForDerivation>,
-      UpdateGet: Class<UpdateGet>,
-    |},
-    IChangePasswordRequest
-  >;
-  +changePrivateDeriverPassword: IChangePasswordRequestFunc,
-}
-
-export type IAddAdhocPublicDeriverRequest<Insert> = AddAdhocPublicDeriverRequest<Insert>;
-export type IAddAdhocPublicDeriverResponse<Row> = AddAdhocPublicDeriverResponse<Row>;
-export type IAddAdhocPublicDeriverFunc<Insert, Row> = (
-  body: IAddAdhocPublicDeriverRequest<Insert>
-) => Promise<IAddAdhocPublicDeriverResponse<Row>>;
-export interface IAdhocPublicDeriver {
-  +rawAddAdhocPubicDeriver: RawTableVariation<
-    IAddAdhocPublicDeriverFunc<mixed, mixed>,
-    {|
-      AddAdhocPublicDeriver: Class<AddAdhocPublicDeriver>,
-    |},
-    IAddAdhocPublicDeriverRequest<mixed>,
-  >;
-  +addAdhocPubicDeriver: IAddAdhocPublicDeriverFunc<mixed, mixed>;
 }
 
 // =====================
